refactor: 고객 banking 공통 모듈 분리
