+++ front/src/lib/customer-banking/types.ts
@@ -1,0 +1,12 @@
+export type MenuSection =
+  | "dashboard"
+  | "accounts"
+  | "transfer"
+  | "transactions"
+  | "notifications"
+  | "security";
+
+export type AlertMessage = {
+  type: "info" | "success" | "error";
+  text: string;
+};
